requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.10,<4.0
python-decouple>=3.5,<4.0
psycopg2>=2.8,<3.0
pillow>=8.0,<9.0
djangorestframework-simplejwt>=4.0,<5.0
redis>=4.0,<5.0
channels>=3.0,<4.0
